feat(MemberReservationRequest): 외부 api 추가 값(paymentKey, orderId, amount) 검증 로직 추가

+++ src/main/java/roomescape/domain/PaymentInfo.java
@@ -2,17 +2,17 @@ package roomescape.domain;
 
 public class PaymentInfo {
 
-    private final int amount;
+    private final Long amount;
     private final String orderId;
     private final String paymentKey;
 
-    public PaymentInfo(int amount, String orderId, String paymentKey) {
+    public PaymentInfo(Long amount, String orderId, String paymentKey) {
         this.amount = amount;
         this.orderId = orderId;
         this.paymentKey = paymentKey;
     }
 
-    public int getAmount() {
+    public Long getAmount() {
         return amount;
     }
